conectado con api qupos

+++ lib/models/product_api.dart
@@ -16,7 +16,7 @@ class ProductApi {
   late double grossWeight;
   late int inStock;
   late int unitType;
-  late int misc1;
+  late double misc1;
   late String misc2;
   late String misc3;
   late String misc4;
@@ -75,7 +75,7 @@ class ProductApi {
         json['unitWeight'] != null ? json['unitWeight'].toDouble() : 0.0;
     grossWeight =
         json['grossWeight'] != null ? json['grossWeight'].toDouble() : 0.0;
-    inStock = json['InStock'] != null ? json['InStock'] : 0;
+    inStock = json['InStock'] != null ? json['InStock'].toInt() : 0;
     unitType = json['UnitType'] != null ? json['UnitType'] : 0;
     misc1 = json['misc1'] ?? 0;
     misc2 = json['misc2'] ?? '';
